+++ scheme/env.go
@@ -1,0 +1,51 @@
+/*
+   Go lang 3rd study program.
+   This is prototype program mini scheme subset.
+
+   hidekuno@gmail.com
+*/
+package scheme
+
+// env structure
+type SimpleEnv struct {
+	EnvTable *Environment
+	Parent   *SimpleEnv
+}
+
+func NewSimpleEnv(parent *SimpleEnv, et *Environment) *SimpleEnv {
+	v := new(SimpleEnv)
+	v.Parent = parent
+	if et != nil {
+		v.EnvTable = et
+	} else {
+		env := Environment{}
+		v.EnvTable = &env
+	}
+	return v
+}
+func (self *SimpleEnv) Find(key string) (Expression, bool) {
+	if v, ok := (*self.EnvTable)[key]; ok {
+		return v, true
+	}
+	for c := self.Parent; c != nil; c = c.Parent {
+		if v, ok := (*c.EnvTable)[key]; ok {
+			return v, true
+		}
+	}
+	return nil, false
+}
+func (self *SimpleEnv) Regist(key string, exp Expression) {
+	(*self.EnvTable)[key] = exp
+}
+func (self *SimpleEnv) Set(key string, exp Expression) {
+	if _, ok := (*self.EnvTable)[key]; ok {
+		(*self.EnvTable)[key] = exp
+		return
+	}
+	for c := self.Parent; c != nil; c = c.Parent {
+		if _, ok := (*c.EnvTable)[key]; ok {
+			(*c.EnvTable)[key] = exp
+			return
+		}
+	}
+}
