+++ scheme/number.go
@@ -1,0 +1,158 @@
+/*
+   Go lang 3rd study program.
+   This is prototype program mini scheme subset.
+
+   hidekuno@gmail.com
+*/
+package scheme
+
+import (
+	"fmt"
+	"reflect"
+	"strconv"
+)
+
+type Number interface {
+	Atom
+	Add(Number) Number
+	Sub(Number) Number
+	Mul(Number) Number
+	Div(Number) Number
+	Equal(Number) bool
+	Greater(Number) bool
+	Less(Number) bool
+	GreaterEqual(Number) bool
+	LessEqual(Number) bool
+}
+
+func CreateNumber(exp Expression) (Number, error) {
+	if v, ok := exp.(*Integer); ok {
+		return NewInteger(v.Value), nil
+	}
+	if v, ok := exp.(*Float); ok {
+		return NewFloat(v.Value), nil
+	}
+	return nil, NewRuntimeError("E1003", reflect.TypeOf(exp).String())
+}
+
+// Integer Type
+type Integer struct {
+	Number
+	Value int
+}
+
+func NewInteger(p int) *Integer {
+	v := new(Integer)
+	v.Value = p
+	return v
+}
+
+func (self *Integer) Add(p Number) Number {
+	v, _ := p.(*Integer)
+	self.Value += v.Value
+	return self
+}
+func (self *Integer) Sub(p Number) Number {
+	v, _ := p.(*Integer)
+	self.Value -= v.Value
+	return self
+}
+func (self *Integer) Mul(p Number) Number {
+	v, _ := p.(*Integer)
+	self.Value *= v.Value
+	return self
+}
+func (self *Integer) Div(p Number) Number {
+	v, _ := p.(*Integer)
+	if v.Value == 0 {
+		panic(NewRuntimeError("E1013"))
+	}
+	self.Value /= v.Value
+	return self
+}
+
+func (self *Integer) Equal(p Number) bool {
+	v, _ := p.(*Integer)
+	return self.Value == v.Value
+}
+func (self *Integer) Greater(p Number) bool {
+	v, _ := p.(*Integer)
+	return self.Value > v.Value
+}
+func (self *Integer) Less(p Number) bool {
+	v, _ := p.(*Integer)
+	return self.Value < v.Value
+}
+func (self *Integer) GreaterEqual(p Number) bool {
+	v, _ := p.(*Integer)
+	return self.Value >= v.Value
+}
+func (self *Integer) LessEqual(p Number) bool {
+	v, _ := p.(*Integer)
+	return self.Value <= v.Value
+}
+
+func (self *Integer) String() string {
+	return strconv.Itoa(self.Value)
+}
+
+// Float Type
+type Float struct {
+	Number
+	Value float64
+}
+
+func NewFloat(p float64) *Float {
+	v := new(Float)
+	v.Value = p
+	return v
+}
+func (self *Float) Add(p Number) Number {
+	v, _ := p.(*Float)
+	self.Value += v.Value
+	return self
+}
+func (self *Float) Sub(p Number) Number {
+	v, _ := p.(*Float)
+	self.Value -= v.Value
+	return self
+}
+func (self *Float) Mul(p Number) Number {
+	v, _ := p.(*Float)
+	self.Value *= v.Value
+	return self
+}
+func (self *Float) Div(p Number) Number {
+	v, _ := p.(*Float)
+	self.Value /= v.Value
+	return self
+}
+func (self *Float) Equal(p Number) bool {
+	v, _ := p.(*Float)
+	return self.Value == v.Value
+}
+func (self *Float) Greater(p Number) bool {
+	v, _ := p.(*Float)
+	return self.Value > v.Value
+}
+func (self *Float) Less(p Number) bool {
+	v, _ := p.(*Float)
+	return self.Value < v.Value
+}
+func (self *Float) GreaterEqual(p Number) bool {
+	v, _ := p.(*Float)
+	return self.Value >= v.Value
+}
+func (self *Float) LessEqual(p Number) bool {
+	v, _ := p.(*Float)
+	return self.Value <= v.Value
+}
+func (self *Float) String() string {
+	return fmt.Sprint(self.Value)
+}
+func (self *Float) FormatString(prec int) string {
+	return strconv.FormatFloat(self.Value, 'f', prec, 64)
+}
+func (self *Float) LogFormatString(prec int) string {
+	return strconv.FormatFloat(self.Value, 'e', prec, 64)
+}
